merge
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/connectomist/constellation-private/trunk@77381 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ src/library/connMatrix.cc
@@ -746,7 +746,7 @@ namespace constel {
   Connectivities * connMatrixReducedFromRegion(
     Connectivities * AllMeshconnMatrixToAllMesh_ptr,
     const TimeTexture<short> & seedRegionsTex, int seedRegionLabel,
-    size_t seedRegionLabelVertexNb,
+    int seedRegionLabelVertexNb,
     vector<size_t> ** seedVertexIndex, bool verbose) {
     if (verbose) cout << "Start connMatrixReducedFromRegion" << endl;
     Connectivities &AllMeshconnMatrixToAllMesh
